Clean up ADT refinement, improve computation notation
The computation notation now inserts linebreaks after [;].  We also
have a tactic for rewriting with the refine monad laws.

The notion of ADT refinement now takes many fewer arguments.  In
particular, we don't need a relation between the old and new models,
just a way of doing the transformation.  The [refine]ment ensures that
we don't break anything.  The way of doing the transformation can even
be non-deterministic, though we may want to remove that feature/bug.
We may also want to make the notion of refinement proof-relevant (so
that we can extract the computation at the very end), or we may not
care, because, after all, the only thing that matters is that the
observers behave the same way.  We should probably prove an end-to-end
theorem that says that, for any initial (new) state, after any
sequence of mutation, all the observers behave the same way as they
would have in the old regime.