Significant step towards a fully functional tree-based refinement
Rewrote the entire tree-based example to use the new nested trees
framework. Observers are fully refined and proved, while some parts of
the mutators are still admitted.

In the process, the dual list/tree structure that the example was
previously using was unified into a single nested tree structure, and
the State inductive types was replaced with nats (though it could
possibly use bounded strings in the future). As a consequence, the list-
based example was removed from the Makefile.