Finish all but one obligation
The remaining one is false.  We have finally reached the point where
we need to minimize the parse tree, rather than taking the top of it.
To do this, I think it suffices to pass around a shallowly-minimal
tree, where, whenever we chop off a name, we recurse down to be sure
we don't see that name until we reset the valid list.